requirements.txt
azure-ai-projects
azure-identity
azure-ai-inference
azure-ai-evaluation
azure-identity
pandas
python-dotenv
jsoncomparison
ipykernel
python-dotenv
